refactor(@sounisi5011/encrypted-archive): rename `src/index.ts` to  `src/index.node.ts`
